# src/resources/npm/api-client/src/models/NoUserExists.ts
/* tslint:disable */
/* eslint-disable */
/**
 * BMLT - OpenAPI 3.0
 * BMLT Admin API Documentation
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * Returns when no user exists.
 * @export
 * @interface NoUserExists
 */
export interface NoUserExists {
    /**
     * 
     * @type {string}
     * @memberof NoUserExists
     */
    message?: string;
}

/**
 * Check if a given object implements the NoUserExists interface.
 */
export function instanceOfNoUserExists(value: object): boolean {
    let isInstance = true;

    return isInstance;
}

export function NoUserExistsFromJSON(json: any): NoUserExists {
    return NoUserExistsFromJSONTyped(json, false);
}

export function NoUserExistsFromJSONTyped(json: any, ignoreDiscriminator: boolean): NoUserExists {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'message': !exists(json, 'message') ? undefined : json['message'],
    };
}

export function NoUserExistsToJSON(value?: NoUserExists | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'message': value.message,
    };
}