# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BMLT - OpenAPI 3.0",
        "description": "BMLT Admin API Documentation",
        "license": {
            "name": "MIT",
            "url": "https://github.com/bmlt-enabled/bmlt-root-server/blob/main/LICENSE"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/users": {
            "get": {
                "tags": [
                    "rootServer"
                ],
                "summary": "Retrieves users",
                "description": "Retrieve users for authenticated user.",
                "operationId": "getUsers",
                "responses": {
                    "200": {
                        "description": "Returns when user is authenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserCollection"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Returns when not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorUnauthenticated"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bmltToken": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "rootServer"
                ],
                "summary": "Creates a user",
                "description": "Creates a user.",
                "operationId": "createUser",
                "requestBody": {
                    "description": "Pass in user object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Returns when POST is successful.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Returns when user is not authenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorIncorrectCredentials"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Returns when user is unauthorized to perform action.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorUnauthenticated"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Returns when no user exists.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NoUserExists"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bmltToken": []
                    }
                ]
            }
        },
        "/api/v1/users/{userId}": {
            "get": {
                "tags": [
                    "rootServer"
                ],
                "summary": "Retrieves a single user",
                "description": "Retrieve single user.",
                "operationId": "getUser",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "ID of user",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns when user is authenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Returns when not authenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorUnauthenticated"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Returns when no user exists.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NoUserExists"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bmltToken": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "rootServer"
                ],
                "summary": "Update single user",
                "description": "Updates a user.",
                "operationId": "updateUser",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "ID of user",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "description": "Pass in user object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Success."
                    },
                    "401": {
                        "description": "Returns when user is not authenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorIncorrectCredentials"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Returns when user is unauthorized to perform action.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorUnauthenticated"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Returns when no user exists.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NoUserExists"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bmltToken": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "rootServer"
                ],
                "summary": "Deletes a user",
                "description": "Deletes a user by id",
                "operationId": "deleteUser",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "ID of user",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Success."
                    },
                    "401": {
                        "description": "Returns when not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorUnauthenticated"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Returns when unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorUnauthorized"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Returns when no user exists.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NoUserExists"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bmltToken": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "rootServer"
                ],
                "summary": "Patches a user",
                "description": "Patches a user by id.",
                "operationId": "partialUpdateUser",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "ID of user",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "description": "Pass in fields you want to update.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserPartialUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Success."
                    },
                    "401": {
                        "description": "Returns when not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorUnauthenticated"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Returns when unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorUnauthorized"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Returns when no user exists.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NoUserExists"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bmltToken": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "ErrorIncorrectCredentials": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The provided credentials are incorrect."
                    }
                },
                "type": "object"
            },
            "ErrorUnauthenticated": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Unauthenticated."
                    }
                },
                "type": "object"
            },
            "ErrorUnauthorized": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "This action is unauthorized."
                    }
                },
                "type": "object"
            },
            "ValidationError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The field is required. (and 1 more error)"
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "example": "error details"
                            }
                        }
                    }
                },
                "type": "object"
            },
...
            "UserBase": {
                "properties": {
                    "username": {
                        "type": "string",
                        "example": "string"
                    },
                    "type": {
                        "type": "string",
                        "example": "string"
                    },
                    "displayName": {
                        "type": "string",
                        "example": "string"
                    },
                    "description": {
                        "type": "string",
                        "example": "string"
                    },
                    "email": {
                        "type": "string",
                        "example": "string"
                    },
                    "ownerId": {
                        "type": "string",
                        "example": "0"
                    }
                },
                "type": "object"
            },
            "User": {
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/UserBase"
                    },
                    {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": "0"
                            }
                        }
                    }
                ]
            },
            "UserCreate": {
                "required": [
                    "username",
                    "password",
                    "type",
                    "displayName"
                ],
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/UserBase"
                    },
                    {
                        "properties": {
                            "password": {
                                "type": "string",
                                "example": "string"
                            }
                        }
                    }
                ]
            },
            "UserUpdate": {
                "required": [
                    "username",
                    "type",
                    "displayName"
                ],
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/UserBase"
                    },
                    {
                        "properties": {
                            "password": {
                                "type": "string",
                                "example": "string"
                            }
                        }
                    }
                ]
            },
            "UserPartialUpdate": {
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/UserBase"
                    },
                    {
                        "properties": {
                            "password": {
                                "type": "string",
                                "example": "string"
                            }
                        }
                    }
                ]
            },
            "UserCollection": {
                "type": "array",
                "items": {
                    "$ref": "#/components/schemas/User"
                }
            },
            "NoUserExists": {
                "description": "Returns when no user exists.",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Not Found"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bmltToken": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "tokenUrl": "api/v1/auth/token",
                        "refreshUrl": "api/v1/auth/refresh",
                        "scopes": {}
                    }
                }
            }
        }
    }
}